+++ app/src/main/java/com/openpositioning/PositionMe/data/remote/WiFiPositioning.java
@@ -1,4 +1,4 @@
-package com.openpositioning.PositionMe.sensors;
+package com.openpositioning.PositionMe.data.remote;
 import android.content.Context;
 import android.util.Log;
 
@@ -89,6 +89,7 @@ public class WiFiPositioning {
                     try {
                             wifiLocation = new LatLng(response.getDouble("lat"),response.getDouble("lon"));
                             floor = response.getInt("floor");
+                            Log.d("WiFiPositioning", "Got WiFi position: " + wifiLocation + ", Floor: " + floor);
                     } catch (JSONException e) {
                         // Error log to keep record of errors (for secure programming and maintainability)
                         Log.e("jsonErrors","Error parsing response: "+e.getMessage()+" "+ response);
