+++ secrets.properties
@@ -1,6 +1,6 @@
 #
 # Modify the variables to set your keys
 #
-MAPS_API_KEY=<your Google maps API key>
-OPENPOSITIONING_API_KEY=<your openpositioning user API key>
-OPENPOSITIONING_MASTER_KEY=<The openpositioning master key>
+MAPS_API_KEY=AIzaSyDDylFZM5nQrcSZ4jE3vOxXw2pcelrfysE
+OPENPOSITIONING_API_KEY=04TZZd3hBEd1k0R_O3XFbQ
+OPENPOSITIONING_MASTER_KEY=ewireless
